Move & Refactor square_bracket tests
Refactor tests to mirror style in other files.
